PIC-1022 entry not available right after failed commit (#260)
transaction roll-back is delayed due to async nature of lease expiration

When transaction lease is known to be expired during a commit,
the thread may exit prematurely on TimeoutExpiredException as the waitFor
timeout during the abort is set to zero.

The abort operation is carried out by the LeaseExpirationMgr, SettlerTask,
and an AbortJob. Each of them is responsible to complete the abort
operation in the background.

Added test to regression:
The test tries to re-take the uncommitted entry and expects to succeed.
Before this fix, due to timing between the abort completion and the re-take,
the entry would return null as if the transaction wasn't yet aborted.

In this timing occurrence it was observed that the abort was completed after
the TimeoutExpiredException was thrown.

Fix:
The fix is to apply the waitFor timeout of the commit operation, waiting for
the abort job to complete. This is the default behavior now. It can be reverted
by disabling with the com.gs.abortNoWaitForExpiredTx=false system property.

The waitFor timeout duration is controlled by the setCommitTimeout.

Co-authored-by: esteratz <ester.atzmon@gigaspaces.com>
Co-authored-by: mavigdor <moran@gigaspaces.com>